poke-mcp/requirements.txt
fastmcp>=2.0.0
python-dotenv>=1.0.0
modal>=0.64.0
